backend/requirements.txt
Django==5.1.6
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.1
django-filter==23.5
Pillow==10.1.0
python-decouple==3.8 